refactor: merge partial loader module into core partial loader
  * move PartialMapLoaderModule declaration/implementation into partial_map_loader hpp/cpp
  * remove obsolete partial_map_loader_module hpp/cpp files
  * update node and test includes to partial_map_loader.hpp
  * add explicit PCL includes in partial loader test after transitive include removal
