+++ map/autoware_map_loader/src/pointcloud_map_loader/partial_map_loader.hpp
@@ -17,9 +17,13 @@
 
 #include "utils.hpp"
 
+#include <rclcpp/rclcpp.hpp>
+
 #include <autoware_map_msgs/msg/area_info.hpp>
+#include <autoware_map_msgs/srv/get_partial_point_cloud_map.hpp>
 
 #include <map>
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -32,6 +36,30 @@ namespace autoware::map_loader
 std::vector<std::string> collect_partial_map_ids(
   const autoware_map_msgs::msg::AreaInfo & area,
   const std::map<std::string, PCDFileMetadata> & pcd_file_metadata_dict);
+
+class PartialMapLoaderModule
+{
+  using GetPartialPointCloudMap = autoware_map_msgs::srv::GetPartialPointCloudMap;
+
+public:
+  explicit PartialMapLoaderModule(std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict);
+
+  // Backward-compatible constructor used by existing tests.
+  explicit PartialMapLoaderModule(
+    rclcpp::Node * node, std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict);
+
+  [[nodiscard]] bool create_response(
+    GetPartialPointCloudMap::Request::SharedPtr req,
+    GetPartialPointCloudMap::Response::SharedPtr res) const;
+
+private:
+  std::map<std::string, PCDFileMetadata> all_pcd_file_metadata_dict_;
+  rclcpp::Service<GetPartialPointCloudMap>::SharedPtr get_partial_pcd_maps_service_;
+
+  [[nodiscard]] bool on_service_get_partial_point_cloud_map(
+    GetPartialPointCloudMap::Request::SharedPtr req,
+    GetPartialPointCloudMap::Response::SharedPtr res) const;
+};
 }  // namespace autoware::map_loader
 
 #endif  // POINTCLOUD_MAP_LOADER__PARTIAL_MAP_LOADER_HPP_
