+++ map/autoware_map_loader/src/pointcloud_map_loader/pointcloud_map_loader_node.cpp
@@ -57,17 +57,6 @@ PointcloudMapLoaderModule::PointcloudMapLoaderModule(
   pub_pointcloud_map_->publish(load_result->loaded_pcd);
 }
 
-PartialMapLoaderModule::PartialMapLoaderModule(
-  rclcpp::Node * node, std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict)
-: PartialMapLoaderModule(std::move(pcd_file_metadata_dict))
-{
-  get_partial_pcd_maps_service_ = node->create_service<GetPartialPointCloudMap>(
-    "service/get_partial_pcd_map",
-    std::bind(
-      &PartialMapLoaderModule::on_service_get_partial_point_cloud_map, this, std::placeholders::_1,
-      std::placeholders::_2));
-}
-
 SelectedMapLoaderModule::SelectedMapLoaderModule(
   rclcpp::Node * node, std::map<std::string, PCDFileMetadata> pcd_file_metadata_dict)
 : SelectedMapLoaderModule(std::move(pcd_file_metadata_dict))
